requirements.txt
aiogram>=3.0.0
openai>=1.0.0
python-dotenv>=1.0.0
